requirements.txt
streamlit>=1.28.0
openslide-python>=1.3.0
Pillow>=10.0.0